[OM-91900] Refactor gitops code to allow multiple remote handlers
